+++ src/models/PeerStatus.ts
@@ -1,0 +1,4 @@
+export interface PeerStatus {
+	isAvailable: boolean;
+	lastStatusCheck: number;
+}